Adds dirty on-token decryption
It works. But there are a lot of traces, original approach I had
in mind isn't good enough. So, next steps would be integration
on-token decryption into UI and changing original concept.
There has to be an option to treat keyfiles as encrypted.
This means that we first decrypt these files using private key stored
on a token. And apply result to keypool.
Also this means that we have to encrypt random data generated for
keyfile with public key, stored on the disk. Or exported from the
token.
And there are issues with the length of data suitable for encryption
and decryption. We have to be aware of this to prepare buffers
of proper length. The questions to answer are "How long our buffer has
to be to store decrypted data, if encrypted data length is N?", "Do we
need to split encrypted data into chunks for PKCS#11 C_Decrypt to
support large buffers?"