Implements security token key selection UI
Added field to mount options dialog. It’s possible to enter key spec in
“slotId:key label” format manually. Or choose key from the key
selection dialog.
Expanded a list of parameters in function, that call
`ApplyListToPassword`. Not of all of them (e.g. password change or
volume generation are not touched). Only of those, that are needed to
mount volume.
Next steps are: figure out what `Keyfiles.c` is used for (Windows
version only?), what’s going on in `CoreBase` (why we call
`ApplyListToPassword` there?), how am I going to generate a new volume
(ideally I need to export public key out of token, generate keyfile as
usual, encrypt first part of it using exported public key using RSA
with default PKCS1 padding), change password (pretty much the same
stuff as in volume creation must be done here).
I might skip all of this, because it works for me. It might be better
to launch VeraCrypt when token is inserted. It should automatically
detect new security token, scan for usable keys, if one is found show
mount volume dialog. Mount volume path and keyfile to use may be
specified beforehand in config file or as a command line option.